requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
Authlib==1.6.3
bcrypt==3.2.2
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.6
curl_cffi==0.11.1
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.12
fastapi-utils==0.8.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
fxcmpy==0.0.0.1
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
numpy==2.2.5
openai==0.28.1
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
paypalrestsdk==1.13.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
psutil==5.9.8
pybit==5.11.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
resend==2.13.1
rich==14.0.0
ruff==0.11.9
safehttpx==0.1.6
schedule==1.2.2
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
stripe==12.4.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
yfinance==0.2.61
